templater/pyproject.toml
[project]
name = "templater"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jinja2>=3.1.6",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.2",
]
